Read the eleven reliability findings; none of them was a defect
Two looked serious and were not: both sat in principalSummary, a helper that builds a log line, not
in the authorisation decision. The check that grants or denies is a plain match over the granted
authorities and is sound. Cleaned anyway, honestly rather than by suppression: the principal is
read once, so the null check and the use are the same value, and the impossible null-authorities
branch is gone.

Two arithmetic warnings, both real patterns and neither reachable: five megabytes fits an int, and
the difference between two image lengths cannot overflow one. Made explicit because the cost is a
character and the next size limit might not be so small.

The rest are shapes rather than faults. The two regexes flagged for catastrophic backtracking are
the safe form of that shape -- each repetition is anchored by a separator, so there is no ambiguity
to explode -- and they validate a product size and a category slug, both typed by staff.

Worth recording because it decides a priority: Reliability C is noise on this codebase today, which
is why the rest of the Sonar backlog can wait behind work that decays.

+++ backend/src/main/java/com/pilarestilo/shared/infrastructure/web/controllers/MediaAdminController.java
@@ -215,7 +215,7 @@ public class MediaAdminController {
                     byte[] original = Files.readAllBytes(file);
                     byte[] optimized = imageOptimizer.reencodeJpeg(original);
                     if (optimized.length < original.length) {
-                        bytesSaved.addAndGet(original.length - optimized.length);
+                        bytesSaved.addAndGet((long) original.length - optimized.length);
                         Files.write(file, optimized);
                     }
                     processed.incrementAndGet();
